requirements.txt
fastapi
uvicorn
python-dotenv
numpy
scipy
openai
